+++ com.unity.render-pipelines.core/Runtime/RenderPipeline/IShaderVariantsSettings.cs
@@ -1,0 +1,28 @@
+namespace UnityEngine.Rendering
+{
+    /// <summary>
+    /// Specifies the logging level for shader variants
+    /// </summary>
+    public enum ShaderVariantLogLevel
+    {
+        Disabled,
+        OnlySRPShaders,
+        AllShaders,
+    }
+
+    /// <summary>
+    /// Exposes settings for shader variants
+    /// </summary>
+    public interface IShaderVariantsSettings
+    {
+        /// <summary>
+        /// Specifies the level of the logging for shader variants
+        /// </summary>
+        ShaderVariantLogLevel shaderVariantLogLevel { get; set; }
+
+        /// <summary>
+        /// Specifies if the stripping of the shaders that variants needs to be exported
+        /// </summary>
+        bool exportShaderVariants { get; set; }
+    }
+}