+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/HDRenderPipelineGlobalSettings.cs
@@ -10,13 +10,6 @@ using UnityEditor;
 
 namespace UnityEngine.Rendering.HighDefinition
 {
-    enum ShaderVariantLogLevel
-    {
-        Disabled,
-        OnlyHDRPShaders,
-        AllShaders,
-    }
-
     enum LensAttenuationMode
     {
         ImperfectLens,
@@ -39,7 +32,7 @@ namespace UnityEngine.Rendering.HighDefinition
     /// - Various resources (such as Shaders) for runtime, editor-only, and raytracing
     /// </summary>
     [HDRPHelpURL("Default-Settings-Window")]
-    partial class HDRenderPipelineGlobalSettings : RenderPipelineGlobalSettings
+    partial class HDRenderPipelineGlobalSettings : RenderPipelineGlobalSettings, IShaderVariantsSettings
     {
         private static HDRenderPipelineGlobalSettings cachedInstance = null;
 
@@ -775,10 +768,6 @@ namespace UnityEngine.Rendering.HighDefinition
         #endregion
 
         #region Misc.
-
-        [SerializeField]
-        internal ShaderVariantLogLevel shaderVariantLogLevel = ShaderVariantLogLevel.Disabled;
-
         [SerializeField]
         internal LensAttenuationMode lensAttenuationMode;
 
@@ -810,6 +799,26 @@ namespace UnityEngine.Rendering.HighDefinition
         }
 
 #endif
+
+        #region IShaderVariantsSettings
+
+        [SerializeField, FormerlySerializedAs("shaderVariantLogLevel")]
+        internal ShaderVariantLogLevel m_ShaderVariantLogLevel = ShaderVariantLogLevel.Disabled;
+
+        /// <summary>
+        /// Specifies the level of the logging for shader variants
+        /// </summary>
+        public ShaderVariantLogLevel shaderVariantLogLevel { get => m_ShaderVariantLogLevel; set => m_ShaderVariantLogLevel = value; }
+
+        [SerializeField]
+        internal bool m_ExportShaderVariants = true;
+
+        /// <summary>
+        /// Specifies if the stripping of the shaders that variants needs to be exported
+        /// </summary>
+        public bool exportShaderVariants { get => m_ExportShaderVariants; set => m_ExportShaderVariants = true; }
+
+        #endregion
 
         [SerializeField]
         internal string DLSSProjectId = "000000";