+++ com.unity.render-pipelines.core/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added DebugFrameTiming class that can be used by render pipelines to display CPU/GPU frame timings and bottlenecks in Rendering Debugger.
 - Added new DebugUI widget types: ProgressBarValue and ValueTuple
 - Added common support code for FSR
+- New extension for Render Pipeline Global Settings for shader variants settings -> `IShaderVariantsSettings`.
 
 ### Fixed
 - Fixed issue when changing volume profiles at runtime with a script (case 1364256).